Enable and document live-reloading in development
* yarn add -D live-server
* yarn add -D watch
* create dev.sh script and ref in pkg json
* add --dev-mode opt to build.sh script
* update README development section

+++ scripts/build.sh
@@ -5,9 +5,16 @@ set -eou pipefail
 # Ensure the script runs from .git root directory
 pushd $(git rev-parse --show-toplevel)
 
-printf "Cleaning dist\n"
-rm -rf dist
-mkdir dist
+if [ -z "${1:-}" ]; then
+  printf "Cleaning dist\n"
+  rm -rf dist
+  mkdir dist
+elif [ "${1}" == "--dev-mode" ]; then
+  printf "Skipping clean\n"
+else
+  printf "Unknown argument ${1}"
+  exit 1
+fi
 
 printf "Copying assets\n"
 cp src/assets/* dist/
@@ -26,12 +33,19 @@ sed -i -e "/%CONTENT%/a $(echo $CONVERTED_TODO_MARKDOWN)" -e "/%CONTENT%/d" dist
 sed -i -e "/%FOOTER%/r src/partials/_footer.html" -e "/%FOOTER%/d" dist/todo.html
 sed -i -e "/\/\/%SCRIPT%/r src/partials/_script.js" -e "/\/\/%SCRIPT%/d" dist/todo.html
 
-printf "Minifying HTML\n"
-./node_modules/.bin/html-minifier --config-file .html-minifier.json --output dist/index.html dist/index.html
-./node_modules/.bin/html-minifier --config-file .html-minifier.json --output dist/todo.html dist/todo.html
+if [ -z "${1:-}" ]; then
+  printf "Minifying HTML\n"
+  ./node_modules/.bin/html-minifier --config-file .html-minifier.json --output dist/index.html dist/index.html
+  ./node_modules/.bin/html-minifier --config-file .html-minifier.json --output dist/todo.html dist/todo.html
 
-printf "Minifying CSS\n"
-./node_modules/.bin/cleancss -O 1 --output dist/style.css dist/style.css
+  printf "Minifying CSS\n"
+  ./node_modules/.bin/cleancss -O 1 --output dist/style.css dist/style.css
+elif [ "${1}" == "--dev-mode" ]; then
+  printf "Skipping minification\n"
+else
+  printf "Unknown argument ${1}"
+  exit 1
+fi
 
 # Reset directory stack
 popd
